Minor Refactor * changed all classes to work relative to the root url. * Class based testing now working. * TODO: Run the server in a way that allows us to test it
