+++ src/aswwu/alchemy.py
@@ -5,9 +5,9 @@ import logging
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, joinedload
-import aswwu.models.bases as base
-import aswwu.models.mask as mask_model
-from aswwu.archive_models import ArchiveBase
+import src.aswwu.models.bases as base
+import src.aswwu.models.mask as mask_model
+from src.aswwu.archive_models import ArchiveBase
 
 Base = base.Base
 ElectionBase = base.ElectionBase
@@ -19,11 +19,11 @@ logger = logging.getLogger("aswwu")
 # import the necessary models (all of them in this case)
 
 # defines the databases URLs relative to "server.py"
-engine = create_engine("sqlite:///../../databases/people.db")
-archive_engine = create_engine("sqlite:///../../databases/archives.db")
-election_engine = create_engine("sqlite:///../../databases/senate_elections.db")
-pages_engine = create_engine("sqlite:///../../databases/pages.db")
-jobs_engine = create_engine("sqlite:///../../databases/jobs.db")
+engine = create_engine("sqlite:///../databases/people.db")
+archive_engine = create_engine("sqlite:///../databases/archives.db")
+election_engine = create_engine("sqlite:///../databases/senate_elections.db")
+pages_engine = create_engine("sqlite:///../databases/pages.db")
+jobs_engine = create_engine("sqlite:///../databases/jobs.db")
 
 # create the model tables if they don't already exist
 Base.metadata.create_all(engine)
