+++ src/aswwu/models/forms.py
@@ -2,7 +2,7 @@ from sqlalchemy import Column, ForeignKey, String, Boolean
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
-import aswwu.models.bases as base
+import src.aswwu.models.bases as base
 
 JobsBase = declarative_base(cls=base.JobsBase)
