+++ src/aswwu/route_handlers/forms.py
@@ -6,9 +6,9 @@ import os
 import bleach
 import tornado.web
 
-from aswwu.base_handlers import BaseHandler
-import aswwu.models.forms as forms_model
-import aswwu.alchemy as alchemy
+from src.aswwu.base_handlers import BaseHandler
+import src.aswwu.models.forms as forms_model
+import src.aswwu.alchemy as alchemy
 
 logger = logging.getLogger("aswwu")
