+++ src/aswwu/route_handlers/pages.py
@@ -3,9 +3,9 @@ import logging
 import bleach
 import tornado.web
 
-from aswwu.base_handlers import BaseHandler
-import aswwu.models.pages as pages_model
-import aswwu.alchemy as alchemy
+from src.aswwu.base_handlers import BaseHandler
+import src.aswwu.models.pages as pages_model
+import src.aswwu.alchemy as alchemy
 
 logger = logging.getLogger("aswwu")
