+++ src/myclass.py
@@ -1,0 +1,4 @@
+# The 
+class myClass:
+    def func(self, x):
+        return x + 1
